feat: backend task endpoints add auth_required decorator

+++ backend/app/routes.py
@@ -1,6 +1,7 @@
 from flask import request, current_app as app
 
-from .models import db, Task, User
+from app.models import db, Task, User
+from app.utils import auth_required
 
 
 @app.route('/')
@@ -37,36 +38,29 @@ def login():
     return {'token': token}, 200
 
 
-@app.route('/test', methods=['GET'])
-def protected():
-    import jwt
-    token = request.headers.get('Authorization')
-    if not token:
-        return {'error': 'Token is missing'}, 401
-    try:
-        user_id = User.decode_auth_token(token)
-        user = User.query.get(user_id)
-        return {'message': f'Welcome {user.username}'}, 200
-    except jwt.ExpiredSignatureError:
-        return {'error': 'Token has expired'}, 401
-    except jwt.InvalidTokenError:
-        return {'error': 'Invalid token'}, 401
+@app.route('/me', methods=['GET'])
+@auth_required
+def get_user(user):
+    return user.serialize(), 200
 
 
 @app.route('/tasks', methods=['GET'])
-def get_all_tasks():
-    tasks = Task.query.all()
+@auth_required
+def get_all_tasks(user):
+    tasks = Task.query.filter_by(user_id=user.id).all()
     return [task.serialize() for task in tasks], 200
 
 
 @app.route('/tasks', methods=['POST'])
-def create_task():
+@auth_required
+def create_task(user):
     data = request.json
     new_task = Task(
         title=data.get('title'),
         description=data.get('description', ''),
         due_date=data.get('due_date'),
         status=data.get('status', Task.STATUS_OPTIONS[0]),
+        user_id=user.id
     )
     errors = new_task.validate()
     if errors:
@@ -77,16 +71,20 @@ def create_task():
 
 
 @app.route('/tasks/<int:task_id>', methods=['GET'])
-def get_task(task_id):
+@auth_required
+def get_task(user, task_id):
     task = Task.query.get(task_id)
     if not task:
         return {'error': f'Task with id {task_id} not found!'}, 404
+    if task.user_id != user.id:
+        return {'error': 'Unauthorized'}, 403
     return task.serialize(), 200
 
 
 @app.route('/tasks/<int:task_id>', methods=['PUT'])
-def update_task(task_id):
-    task = Task.query.get(task_id)
+@auth_required
+def update_task(user, task_id):
+    task = Task.query.filter(user_id=user.id, id=task_id).first()
     if not task:
         return {'error': f'Task with id {task_id} not found!'}, 404
     data = request.json
@@ -99,10 +97,13 @@ def update_task(task_id):
 
 
 @app.route('/tasks/<int:task_id>', methods=['DELETE'])
-def delete_task(task_id):
+@auth_required
+def delete_task(user, task_id):
     task = Task.query.get(task_id)
     if not task:
         return {'error': f'Task with id {task_id} not found!'}, 404
+    if task.user_id != user.id:
+        return {'error': 'Unauthorized'}, 403
     db.session.delete(task)
     db.session.commit()
     return {}, 204
